requirements.txt
agate==1.9.1
altair==5.4.1
annotated-types==0.7.0
appdirs==1.4.4
asttokens==2.4.1
attrs==24.2.0
babel==2.16.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
daff==1.3.46
dbt-adapters==1.4.1
dbt-common==1.7.0
dbt-core==1.8.6
dbt-extractor==0.5.1
dbt-postgres==1.8.2
dbt-semantic-interfaces==0.5.1
decorator==5.1.1
deepdiff==7.0.1
defusedxml==0.7.1
delta-spark==3.2.0
diff_cover==9.1.1
docopt==0.6.2
exceptiongroup==1.2.2
executing==2.1.0
fastjsonschema==2.20.0
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.1.1
idna==3.8
importlib-metadata==6.11.0
iniconfig==2.0.0
ipython==8.12.3
isodate==0.6.1
jedi==0.19.1
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
leather==0.4.0
Logbook==1.5.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mashumaro==3.13.1
matplotlib-inline==0.1.7
mdurl==0.1.2
minimal-snowplow-tracker==0.0.2
mistune==3.0.2
more-itertools==10.5.0
msgpack==1.0.8
narwhals==1.8.4
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
networkx==3.3
numpy==2.1.1
ordered-set==4.1.0
packaging==24.1
pandas==2.2.3
pandocfilters==1.5.1
parsedatetime==2.6
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pipreqs==0.5.0
platformdirs==4.3.6
pluggy==1.5.0
prompt_toolkit==3.0.48
protobuf==4.25.4
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
py4j==0.10.9.7
pyarrow==17.0.0
pydantic==2.9.0
pydantic_core==2.23.2
pydeck==0.9.1
Pygments==2.18.0
pyspark==3.5.2
pytest==8.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pytimeparse==1.1.8
pytz==2024.1
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
rich==13.8.1
rpds-py==0.20.0
six==1.16.0
smmap==5.0.1
soupsieve==2.6
SQLAlchemy==2.0.35
sqlfluff==3.1.1
sqlparse==0.5.1
stack-data==0.6.3
streamlit==1.38.0
tblib==3.0.0
tenacity==8.5.0
text-unidecode==1.3
tinycss2==1.3.0
toml==0.10.2
tomli==2.0.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
watchdog==4.0.2
wcwidth==0.2.13
webencodings==0.5.1
yarg==0.1.9
zipp==3.20.1
